+++ src/avt/VisWindow/Colleagues/avtText2DColleague.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -551,7 +551,7 @@ avtText2DColleague::SetText(const char *formatString)
     // Save the format string. Don't do it in the case that the formatString
     // pointer is the same as textFormatString, which is how we get here from
     // UpdatePlotList.
-    int len = strlen(formatString);
+    size_t len = strlen(formatString);
     if(textFormatString != formatString)
     {
         delete [] textFormatString;
@@ -565,7 +565,7 @@ avtText2DColleague::SetText(const char *formatString)
     std::string::size_type pos;
     if((pos=fmtStr.find(TIME_IDENTIFIER)) != std::string::npos)
     {
-        int tlen = strlen(TIME_IDENTIFIER);
+        size_t tlen = strlen(TIME_IDENTIFIER);
         std::string left(fmtStr.substr(0, pos));
         std::string right(fmtStr.substr(pos + tlen, fmtStr.size() - pos - tlen));
         char tmp[100];
@@ -576,7 +576,7 @@ avtText2DColleague::SetText(const char *formatString)
     }
     else if((pos=fmtStr.find(CYCLE_IDENTIFIER)) != std::string::npos)
     {
-        int tlen = strlen(CYCLE_IDENTIFIER);
+        size_t tlen = strlen(CYCLE_IDENTIFIER);
         std::string left(fmtStr.substr(0, pos));
         std::string right(fmtStr.substr(pos + tlen, fmtStr.size() - pos - tlen));
         char tmp[100];
