+++ src/common/misc/VisItInit.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -653,7 +653,7 @@ RemovePrependedDirs(const char *path, char *name)
     //
     // Find the last slash by going to the end and working backwards.
     //
-    int  len = strlen(path);
+    int  len = (int)strlen(path);
     int lastSlash;
     for (lastSlash=len ; lastSlash>=0 && path[lastSlash]!=visitcommon::SlashChar() ; lastSlash--)
     {
