+++ src/common/state/ClientInformationList.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -573,7 +573,7 @@ ClientInformationList::RemoveClients(int index)
 int
 ClientInformationList::GetNumClients() const
 {
-    return clients.size();
+    return (int)clients.size();
 }
 
 // ****************************************************************************
