+++ src/databases/OVERFLOW/avtOVERFLOWFileFormat.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -113,7 +113,7 @@ avtOVERFLOWFileFormat::InitializeFile()
     // Find the last slash -- indicates the first character of the filename
     //
     int lastslash;
-    for (lastslash=origfilename.size()-2; lastslash>0; lastslash--)
+    for (lastslash=(int)origfilename.size()-2; lastslash>0; lastslash--)
     {
         if (origfilename[lastslash]=='/')
             break;
