+++ src/databases/SAS/avtSASFileFormat.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -221,7 +221,7 @@ avtSASFileFormat::GetNTimesteps(void)
     if (aTimes.size() == 0)
         ReadTimeStepData();
 
-    return aTimes.size();
+    return (int)aTimes.size();
 }
