+++ src/databases/SimV2/avtSimV2FileFormat.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -499,7 +499,7 @@ RestrictMaterialIndices(visit_handle h, const avtMaterialMetaData *mmd)
                     {
                         if(mmd->materialNames[idx] == std::string(matname))
                         {
-                            restrictedMats.push_back(idx);
+                            restrictedMats.push_back((int)idx);
                             break;
                         }
                     }
