+++ src/gui/NameSimplifier.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2012, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -158,7 +158,7 @@ NameSimplifier::UniqueFileName::UniqueFileName(const QualifiedFilename &qfn)
     separator = qfn.separator;
     leadingSlash = (qfn.path.size() > 0 && qfn.path[0] == separator);
     path = SplitValues(qfn.path, separator);
-    pathLen = path.size();
+    pathLen = (int)path.size();
 }
